# internal/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "csjk backend",
        "title": "csjk-bk",
        "contact": {
            "email": "hecheng@nscc-tj.cn"
        },
        "version": "0.0.1-alpha"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/:cluster/ldap/group": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理",
                    "用户管理"
                ],
                "summary": "在某集群 ldap 中创建用户组",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"test\"",
                        "description": "集群名称",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/:cluster/ldap/group/:name": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理",
                    "用户管理"
                ],
                "summary": "在某集群 ldap 中更新用户组信息",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"test\"",
                        "description": "集群名称",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户组名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理",
                    "用户管理"
                ],
                "summary": "在某集群 ldap 中删除用户组",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"test\"",
                        "description": "集群名称",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户组名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/:cluster/ldap/group/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理",
                    "用户管理"
                ],
                "summary": "获取 ldap 用户组列表",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"test\"",
                        "description": "集群名称",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "是否分页",
                        "name": "paging",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ldap.GroupListItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/:cluster/ldap/user": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理",
                    "用户管理"
                ],
                "summary": "在某集群 ldap 中创建用户",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"test\"",
                        "description": "集群名称",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/:cluster/ldap/user/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理",
                    "用户管理"
                ],
                "summary": "获取某集群 LDAP 用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"test\"",
                        "description": "集群名称",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "是否分页",
                        "name": "paging",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ldap.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ldap.GroupListItem": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "ldap.description",
                    "type": "string"
                },
                "gid": {
                    "description": "ldap.GidNumber",
                    "type": "integer"
                },
                "name": {
                    "description": "ldap.cn",
                    "type": "string"
                },
                "users": {
                    "description": "ldap.memberuid",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "results": {}
            }
        },
...
    }
}